requirements.txt
langchain_community
langchain_text_splitters
pypdf
python-docx
langchain-chroma
sentence-transformers
fastapi
uvicorn
python-multipart
